new mac build, new libs, xcode etc

+++ src/Utility/gl/GLUtils.cpp
@@ -170,7 +170,7 @@ void GLUtils::initGL(char* windowName)
 	//set up display mode
 	//-----------------------------
 
-	Uint32 start, now, totalStart, totalNow;
+	Uint64 start, now, totalStart, totalNow;
 	start = SDL_GetPerformanceCounter();
 	totalStart = SDL_GetPerformanceCounter();
